# app/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
paths:
...
  /bookmarks:
    post:
      summary: Toggle a job bookmark
      description: Adds or removes a bookmark for the authenticated user.
      tags:
        - Bookmarks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                job_id:
                  type: integer
                  description: ID of the job to bookmark or remove.
                  example: 123
      responses:
        '201':
          description: Bookmark added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Bookmark added.
        '200':
          description: Bookmark removed successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Bookmark removed.
        '400':
          description: Job ID is missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Job ID is required.
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error.

    get:
      summary: Get a list of bookmarks
      description: Retrieves the list of jobs bookmarked by a user, sorted by the latest bookmarks first, with pagination.
      tags:
        - Bookmarks
      security:
        - BearerAuth: [ ]
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          required: false
          schema:
            type: integer
          example: 1
        - name: per_page
          in: query
          description: Number of items per page.
          required: false
          schema:
            type: integer
          example: 20
      responses:
        '200':
          description: List of bookmarks retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        bookmark_id:
                          type: integer
                          example: 1
                        user_id:
                          type: integer
                          example: 123
                        job_id:
                          type: integer
                          example: 456
                        title:
                          type: string
                          example: Backend Developer
                        company_name:
                          type: string
                          example: ABC Corp
                        salary:
                          type: string
                          example: $60,000
                        deadline:
                          type: string
                          format: date
                          example: 2024-12-31
                  pagination:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 20
                      total_items:
                        type: integer
                        example: 50
        '400':
          description: Missing or invalid query parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: User ID is required.
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Internal server error.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT